model_server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
llama-cpp-python>=0.2.87
